QLMatHang

+++ QuanLySanBong/Model/ChiTietHoaDonBan.cs
@@ -12,7 +12,7 @@ namespace QuanLySanBong.Model
         [Key]
         [Column(Order = 0)]
         [DatabaseGenerated(DatabaseGeneratedOption.None)]
-        public long IDHoaDon { get; set; }
+        public int IDHoaDon { get; set; }
 
         [Key]
         [Column(Order = 1)]
